Controlamos sesion en HomeScreen, si no esta logueado enviamos a signup
